requirements-dev.txt
coveralls
pytest
pytest-cov
pytest-xdist
pytest-django
flake8
fabric
mock
responses
